@@ -1,9 +1,9 @@
 ﻿using Coreflow.Interfaces;
 using Coreflow.Validation.Messages;
-using Newtonsoft.Json.Linq;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Text.Json;
 
 namespace Coreflow.Validation.Corrector
 {
@@ -42,7 +42,7 @@ namespace Coreflow.Validation.Corrector
 
         public static void Correct(FlowDefinition pFlowDefinition, List<Guid> pCodeCreators, object pData)
         {
-            var data = ((JObject)pData).ToObject<MoveDataContainer>();
+            var data = JsonSerializer.Deserialize<MoveDataContainer>(pData.ToString()); // ((JObject)pData).ToObject<MoveDataContainer>();
 
             foreach (var ccid in pCodeCreators)
             {
